Move layout to trait
