+++ src/Fields/Checkbox.php
@@ -13,10 +13,10 @@ declare(strict_types=1);
 
 namespace WordPlate\Acf\Fields;
 
-use InvalidArgumentException;
 use WordPlate\Acf\Fields\Attributes\Choices;
 use WordPlate\Acf\Fields\Attributes\ConditionalLogic;
 use WordPlate\Acf\Fields\Attributes\DefaultValue;
+use WordPlate\Acf\Fields\Attributes\DirectionLayout;
 use WordPlate\Acf\Fields\Attributes\Instructions;
 use WordPlate\Acf\Fields\Attributes\Required;
 use WordPlate\Acf\Fields\Attributes\ReturnFormat;
@@ -26,6 +26,7 @@ class Checkbox extends Field
 {
     use Choices;
     use DefaultValue;
+    use DirectionLayout;
     use ConditionalLogic;
     use Instructions;
     use Required;
@@ -33,15 +34,4 @@ class Checkbox extends Field
     use Wrapper;
 
     protected $type = 'checkbox';
-
-    public function layout(string $layout): self
-    {
-        if (!in_array($layout, ['vertical', 'horizontal'])) {
-            throw new InvalidArgumentException("Invalid argument layout [$layout].");
-        }
-
-        $this->config->set('layout', $layout);
-
-        return $this;
-    }
 }
